Refactor routes to reuse navbar and sidebar across pages
