+++ src/components/common/stat_widget.js
