+++ src/components/layouts/navigation_bar.js
@@ -5,6 +5,7 @@ import { Link } from 'react-router';
 export default class NavBar extends Component {
   render() {
     return (
+      <div>
       <Navbar fluid>
         <Navbar.Header>
           <Navbar.Brand>
@@ -141,8 +142,11 @@ export default class NavBar extends Component {
             </ul>
           </div>
         </div>
-
       </Navbar>
+
+      {this.props.children}
+
+      </div>
     );
   }
 }
