+++ src/components/pages/indicators.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import MyPageHeader from '../common/page_header';
+
+const Indicators = () => {
+    return (
+      <div className="page-wrapper content">
+        <MyPageHeader title="Indicators" />
+      </div>
+    );
+};
+
+export default Indicators;
