+++ src/routes.js
@@ -1,18 +1,20 @@
 import React from 'react';
-import { Router, Route } from 'react-router';
+import { IndexRoute, Route } from 'react-router';
 
-import App from './components/app';
-import Accounts from './components/accounts';
-import Balances from './components/cash_balances';
-import Indicators from './components/indicators';
-import ReferenceData from './components/ref_data';
+import NavLayout from "./components/layouts/navigation_bar";
+import HomePage from "./components/pages/dashboard";
+import Balances from "./components/pages/cash_balances";
+import Accounts from "./components/pages/accounts";
+import Indicators from "./components/pages/indicators";
+import ReferenceData from "./components/pages/ref_data";
 
 export default (
-  <Router>
-    <Route path="/" component={App} />
-    <Route path="/accounts" component={Accounts} />
-    <Route path="/balances" component={Balances} />
-    <Route path="/indicators" component={Indicators} />
-    <Route path="/ref_data" component={ReferenceData} />
-  </Router>
+    <Route path="/" component={NavLayout}>
+      <IndexRoute component={HomePage} />
+      <Route path="/dashboard" component={HomePage} />
+      <Route path="/balances" component={Balances} />
+      <Route path="/indicators" component={Indicators} />
+      <Route path="/accounts" component={Accounts} />
+      <Route path="/ref_data" component={ReferenceData} />
+    </Route>
 );
